refactor: digest path
